refactor(desktop): organize home hooks
