+++ desktop/src/hooks/home/derived/use-home-next-mode-selection.ts
