+++ desktop/src/hooks/home/derived/use-home-next-model-selection.test.tsx
